+++ frontend/src/features/Mission/components/MissionForm/ActionForm/shared/utils.ts
@@ -1,6 +1,8 @@
+import { EU_COUNTRY_CODES } from '@features/Alert/components/SideWindowAlerts/AlertManagementForm/constants'
 import { MissionAction } from '@features/Mission/missionAction.types'
-import { getLocalizedDayjs } from '@mtes-mct/monitor-ui'
+import { getLocalizedDayjs, pluralize } from '@mtes-mct/monitor-ui'
 
+import type { MissionActionFormValues } from '../../types'
 import type { FleetSegment } from '@features/FleetSegment/types'
 import type { Option } from '@mtes-mct/monitor-ui'
 
@@ -46,5 +48,53 @@ export function getNumberInFrench(aNumber: number): string {
       return 'neuf'
     default:
       return String(aNumber)
+  }
+}
+
+export type PriorityTargetReasons = {
+  hasGroupOrReportingReason: boolean
+  isPriorityTarget: boolean
+  reasons: string[]
+}
+
+export function getPriorityTargetReasons(
+  values: Pick<
+    MissionActionFormValues,
+    'actionType' | 'flagState' | 'isINNControl' | 'portLocode' | 'tripReportings' | 'vesselGroups'
+  >
+): PriorityTargetReasons {
+  const priorityGroups = (values.vesselGroups ?? []).filter(group => group.isPriorityGroup)
+  const currentTripReportingLength = (values.tripReportings ?? []).length
+  const hasGroupOrReportingReason = priorityGroups.length > 0 || currentTripReportingLength > 0
+  const isThirdCountryVesselLandingInFrance =
+    values.actionType === MissionAction.MissionActionType.LAND_CONTROL &&
+    !!values.portLocode?.startsWith('FR') &&
+    !!values.flagState &&
+    values.flagState !== 'UNDEFINED' &&
+    !EU_COUNTRY_CODES.includes(values.flagState)
+
+  const reasons: string[] = []
+  if (priorityGroups.length > 0) {
+    const groupNames = priorityGroups.map(group => `“${group.name}”`).join(' et ')
+    reasons.push(
+      `il appartient au${priorityGroups.length > 1 ? 'x' : ''} ${pluralize('groupe', priorityGroups.length)} ${pluralize('prioritaire', priorityGroups.length)} ${groupNames}`
+    )
+  }
+  if (currentTripReportingLength > 0) {
+    reasons.push(
+      `${getNumberInFrench(currentTripReportingLength)} ${pluralize('suspicion', currentTripReportingLength)} d’infraction est en cours sur sa marée`
+    )
+  }
+  if (values.isINNControl) {
+    reasons.push('c’est un navire INN')
+  }
+  if (isThirdCountryVesselLandingInFrance) {
+    reasons.push('c’est un navire tiers débarquant dans un port français')
+  }
+
+  return {
+    hasGroupOrReportingReason,
+    isPriorityTarget: hasGroupOrReportingReason || !!values.isINNControl || isThirdCountryVesselLandingInFrance,
+    reasons
   }
 }
